update: ahci read/write size
now we can read/write with max size

+++ includes/driver/ahci.h
@@ -700,8 +700,8 @@ typedef struct ahci_hba_s {
 
 int8_t   ahci_init(memory_heap_t* heap, linkedlist_t sata_pci_devices);
 int8_t   ahci_identify(uint64_t disk_id);
-future_t ahci_read(uint64_t disk_id, uint64_t lba, uint16_t size, uint8_t* buffer);
-future_t ahci_write(uint64_t disk_id, uint64_t lba, uint16_t size, uint8_t* buffer);
+future_t ahci_read(uint64_t disk_id, uint64_t lba, uint32_t size, uint8_t* buffer);
+future_t ahci_write(uint64_t disk_id, uint64_t lba, uint32_t size, uint8_t* buffer);
 future_t ahci_flush(uint64_t disk_id);
 
 const ahci_sata_disk_t* ahci_get_disk_by_id(uint64_t disk_id);
